22.0.0_240215

+++ include/win32/jni_md.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1996, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1996, 2023, Oracle and/or its affiliates. All rights reserved.
  * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
  *
  *
@@ -32,9 +32,8 @@
 #define JNIIMPORT __declspec(dllimport)
 #define JNICALL __stdcall
 
-// 'long' is always 32 bit on windows so this matches what jdk expects
-typedef long jint;
-typedef __int64 jlong;
+typedef int jint;
+typedef long long jlong;
 typedef signed char jbyte;
 
 #endif /* !_JAVASOFT_JNI_MD_H_ */
